refactor(cli): split main.py into per-workspace command modules
PR-1 of the unified-IA spec. The 4132-line main.py is split into 8 per-workspace command modules plus a slim main.py. Mechanical extraction — every command name, signature, body, and registered-with-name is preserved byte-identical. Zero behavior change.

Layout:
  cli/forge/
    main.py                 # 126 lines — root app + register() wiring
    commands/
      __init__.py
      system.py             # version, init, up, down, restart, status, dashboard, costs, health + lifecycle helpers
      auth.py               # auth_app + flat whoami / login / logout
      studio.py             # agents_app, blueprints_app, prompts_app, knowledge_app, workspace_app
      ops.py                # runs_app, orchestrate cmd + orchestrate_app, approvals_app, triggers_app, traces_app, recordings_app, messages_app (+ mail), trace cmd
      evals.py              # evals_app + flat compare cmd
      connections.py        # models_app, mcp_app, targets_app, cu_app (+ nested backends_app), tools_app
      marketplace.py        # marketplace_app
      settings.py           # teams_app, keys_app, config_app

Module shape (consistent across all eight):
  - Imports + the original main.py prelude (helper imports, PIDS_FILE, console).
  - `_app = typer.Typer()` — captures every `@app.command()`-decorated flat command in the module.
  - The original sub-app `typer.Typer(...)` declarations + their `@<sub>_app.command()` bodies, untouched.
  - `register(parent)` forwards `_app.registered_commands` onto `parent` so flat commands stay at the root, then calls `parent.add_typer(<sub_app>, name="...")` for each sub-app — same names that main.py used to register directly.

Acceptance criteria:
  - `forge --help` output is identical in command set to before (verified manually — all 15 flat commands + 25 group registrations including the existing computer-use/cu and messages/mail dual aliases).
  - All existing CLI tests pass untouched (`cli/tests/test_cli.py` 7/7 + `cli/tests/test_config_round_trip.py` 1/1).
  - main.py is 126 lines (spec target was < 300).

Notes:
  - workspace_app (file-management subcommands for agent runtimes) lives in studio.py — it's an agent-execution surface, not in the spec's explicit mapping but the closest fit.
  - backends_app is defined in connections.py but mounted under cu_app, exactly as in the original.
  - ruff auto-fix cleaned 133 of the duplicated-import findings the split introduced; the remaining 152 are pre-existing unused-arg warnings carried over verbatim from the original file.

PR-2 (re-parenting into the six workspaces with back-compat aliases) builds on this structure.
